Migrate Realtime connection Resume tests
