--- test/realtime/resume.js
@@ -1,4 +1,0 @@
-"use strict";
-var base = require('./common/common').setup();
-var mod = require('./common/resume');
-base.addCommonModule(exports, mod);
